Crate bridge_runtime_common

Common types/functions that may be used by runtimes of all bridged chains.

Modules

extensions

Bridge-specific transaction extensions.

messages_api

Helpers for implementing various message-related runtime API methods.

messages_call_ext

Helpers for easier manipulation of call processing with signed extensions.

messages_xcm_extension

Module provides utilities for easier XCM handling, e.g: XcmExecutor -> MessageSender -> OutboundMessageQueue | Relayer | XcmRouter <- MessageDispatch <- InboundMessageQueue

Macros

generate_bridge_reject_obsolete_headers_and_messages

Declares a runtime-specific BridgeRejectObsoleteHeadersAndMessages signed extension.