bridge_runtime_common

Module extensions

Transaction extension that rejects bridge-related transactions, that include obsolete (duplicated) data or do not pass some additional pallet-specific checks.

Structs

CheckAndBoostBridgeGrandpaTransactions

Wrapper for the bridge GRANDPA pallet that checks calls for obsolete submissions and also boosts transaction priority if it has submitted by registered relayer. The boost is computed as (BundledHeaderNumber - 1 - BestFinalizedHeaderNumber) * Priority::get(). The boost is only applied if submitter has active registration in the relayers pallet.

CheckAndBoostBridgeParachainsTransactions

Wrapper for the bridge parachains pallet that checks calls for obsolete submissions and also boosts transaction priority if it has submitted by registered relayer. The boost is computed as (BundledHeaderNumber - 1 - BestKnownHeaderNumber) * Priority::get(). The boost is only applied if submitter has active registration in the relayers pallet.

Traits

BridgeRuntimeFilterCall

A duplication of the FilterCall trait.