Module bridge_runtime_common::extensions::priority_calculator

Bridge transaction priority calculator.

We want to prioritize message delivery transactions with more messages over transactions with less messages. That’s because we reject delivery transactions if it contains already delivered message. And if some transaction delivers single message with nonce N, then the transaction with nonces N..=N+100 will be rejected. This can lower bridge throughput down to one message per block.

Functions

compute_priority_boost

Compute priority boost for transaction that brings given number of bridge items (messages, headers, …), when every additional item adds PriorityBoostPerItem to transaction priority.

Type Aliases

ItemCount

We’ll deal with different bridge items here - messages, headers, … To avoid being too verbose with generic code, let’s just define a separate alias.