Module bridge_runtime_common::extensions::refund_relayer_extension

Signed extension that refunds relayer if he has delivered some new messages. It also refunds transaction cost if the transaction is an utility.batchAll() with calls that are: delivering new message and all necessary underlying headers (parachain or relay chain).

Structs

ActualFeeRefund

RefundCalculator implementation which refunds the actual transaction fee.

PreDispatchData

Data that is crafted in pre_dispatch method and used at post_dispatch.

RefundBridgedGrandpaMessages

Signed extension that refunds a relayer for new messages coming from a standalone (GRANDPA) chain.

RefundBridgedMessages

Transaction extension that refunds a relayer for standalone messages delivery and confirmation transactions. Finality transactions are not refunded.

RefundBridgedParachainMessages

Signed extension that refunds a relayer for new messages coming from a parachain.

RefundSignedExtensionAdapter

Adapter that allow implementing sp_runtime::traits::SignedExtension for any RefundSignedExtension.

RefundableMessagesLane

Default implementation of RefundableMessagesLaneId.

RefundableParachain

Implementation of RefundableParachainId for trait Parachain.

Enums

CallInfo

Type of the call that the extension recognizes.

RelayerAccountAction

The actions on relayer account that need to be performed because of his actions.

Traits

RefundCalculator

Refund calculator.

RefundSignedExtension

Everything common among our refund signed extensions.

RefundableMessagesLaneId

Trait identifying a bridged messages lane. A relayer might be refunded for delivering messages coming from this lane.

RefundableParachainId

Trait identifying a bridged parachain. A relayer might be refunded for delivering messages coming from this parachain.