bridge_runtime_common::extensions

Trait BridgeRuntimeFilterCall

pub trait BridgeRuntimeFilterCall<AccountId, Call> {
    type ToPostDispatch;

    // Required method
    fn validate(
        who: &AccountId,
        call: &Call,
    ) -> (Self::ToPostDispatch, TransactionValidity);

    // Provided method
    fn post_dispatch(
        _who: &AccountId,
        _has_failed: bool,
        _to_post_dispatch: Self::ToPostDispatch,
    ) { ... }
}

A duplication of the FilterCall trait.

We need this trait in order to be able to implement it for the messages pallet, since the implementation is done outside of the pallet crate.

Required Associated Types

type ToPostDispatch

Data that may be passed from the validate to post_dispatch.

Required Methods

fn validate( who: &AccountId, call: &Call, ) -> (Self::ToPostDispatch, TransactionValidity)

Called during validation. Needs to checks whether a runtime call, submitted by the who is valid. Transactions not signed are not validated.

Provided Methods

fn post_dispatch( _who: &AccountId, _has_failed: bool, _to_post_dispatch: Self::ToPostDispatch, )

Called after transaction is dispatched.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.