bridge_runtime_common

Macro generate_bridge_reject_obsolete_headers_and_messages

macro_rules! generate_bridge_reject_obsolete_headers_and_messages {
    ($call:ty, $account_id:ty, $($filter_call:ty),*) => { ... };
}

Declares a runtime-specific BridgeRejectObsoleteHeadersAndMessages signed extension.

Example

generate_bridge_reject_obsolete_headers_and_messages!{
    Call, AccountId
    BridgeRococoGrandpa, BridgeRococoMessages,
    BridgeRococoParachains
}

The goal of this extension is to avoid “mining” transactions that provide outdated bridged headers and messages. Without that extension, even honest relayers may lose their funds if there are multiple relays running and submitting the same information.