Function bridge_runtime_common::messages::transaction_payment

pub fn transaction_payment<Balance: AtLeast32BitUnsigned + FixedPointOperand>(
    base_extrinsic_weight: Weight, 
    per_byte_fee: Balance, 
    multiplier: FixedU128, 
    weight_to_fee: impl Fn(Weight) -> Balance, 
    transaction: MessageTransaction<Weight>
) -> Balance

Compute fee of transaction at runtime where regular transaction payment pallet is being used.

The value of multiplier parameter is the expected value of pallet_transaction_payment::NextFeeMultiplier at the moment when transaction is submitted. If you’re charging this payment in advance (and that’s what happens with delivery and confirmation transaction in this crate), then there’s a chance that the actual fee will be larger than what is paid in advance. So the value must be chosen carefully.