Module bridge_runtime_common::messages::source

Sub-module that is declaring types required for processing This -> Bridged chain messages.

Structs

FromBridgedChainMessagesDeliveryProof

Messages delivery proof from bridged chain:

FromThisChainMessageVerifier

Message verifier that is doing all basic checks.

Constants

BAD_ORIGIN

The error message returned from LaneMessageVerifier when call origin is mismatch.

OUTBOUND_LANE_DISABLED

The error message returned from LaneMessageVerifier when outbound lane is disabled.

TOO_LOW_FEE

The error message returned from LaneMessageVerifier when the message fee is too low.

TOO_MANY_PENDING_MESSAGES

The error message returned from LaneMessageVerifier when too many pending messages at the lane.

Functions

estimate_message_dispatch_and_delivery_fee

Estimate delivery and dispatch fee that must be paid for delivering a message to the Bridged chain.

maximal_message_size

Return maximal message size of This -> Bridged chain message.

verify_chain_message

Do basic Bridged-chain specific verification of This -> Bridged chain message.

verify_messages_delivery_proof

Verify proof of This -> Bridged chain messages delivery.

Type Definitions

BridgedChainOpaqueCall

Encoded Call of the Bridged chain. We never try to decode it on This chain.

FromThisChainMessagePayload

Message payload for This -> Bridged chain messages.

ParsedMessagesDeliveryProofFromBridgedChain

‘Parsed’ message delivery proof - inbound lane id and its state.