Module bridge_runtime_common::messages::target

Sub-module that is declaring types required for processing Bridged -> This chain messages.

Structs

FromBridgedChainEncodedMessageCall

Encoded Call of This chain as it is transferred over bridge.

FromBridgedChainMessageDispatch

Dispatching Bridged -> This chain messages.

FromBridgedChainMessagesProof

Messages proof from bridged chain:

Functions

maximal_incoming_message_dispatch_weight

Return maximal dispatch weight of the message we’re able to receive.

maximal_incoming_message_size

Return maximal message size given maximal extrinsic size.

verify_messages_proof

Verify proof of Bridged -> This chain messages.

Type Definitions

FromBridgedChainMessageCallOrigin

Call origin for Bridged -> This chain messages.

FromBridgedChainMessagePayload

Decoded Bridged -> This message payload.