Trait bridge_runtime_common::messages::BridgedChainWithMessages

pub trait BridgedChainWithMessages: ChainWithMessages {
    fn maximal_extrinsic_size() -> u32;

    fn message_weight_limits(
        message_payload: &[u8]
    ) -> RangeInclusive<Self::Weight>;

    fn estimate_delivery_transaction(
        message_payload: &[u8], 
        include_pay_dispatch_fee_cost: bool, 
        message_dispatch_weight: WeightOf<Self>
    ) -> MessageTransaction<WeightOf<Self>>;

    fn transaction_payment(
        transaction: MessageTransaction<WeightOf<Self>>
    ) -> BalanceOf<Self>;
}

Bridged chain that has pallet-bridge-messages and dispatch modules.

Required methods

fn maximal_extrinsic_size() -> u32

Maximal extrinsic size at Bridged chain.

fn message_weight_limits(message_payload: &[u8]) -> RangeInclusive<Self::Weight>

Returns feasible weights range for given message payload at the Bridged chain.

If message is being sent with the weight that is out of this range, then it should be rejected.

Weights returned from this function shall not include transaction overhead (like weight of signature and signed extensions verification), because they’re already accounted by the weight_of_delivery_transaction. So this function should return pure call dispatch weights range.

fn estimate_delivery_transaction(
message_payload: &[u8],
include_pay_dispatch_fee_cost: bool,
message_dispatch_weight: WeightOf<Self>
) -> MessageTransaction<WeightOf<Self>>

Estimate size and weight of single message delivery transaction at the Bridged chain.

fn transaction_payment(
transaction: MessageTransaction<WeightOf<Self>>
) -> BalanceOf<Self>

Returns minimal transaction fee that must be paid for given transaction at the Bridged chain.