Trait bridge_runtime_common::messages::ChainWithMessages

pub trait ChainWithMessages {
    type Hash: Decode;
    type AccountId: Encode + Decode;
    type Signer: Encode + Decode;
    type Signature: Encode + Decode;
    type Weight: From<Weight> + PartialOrd;
    type Balance: Encode + Decode + CheckedAdd + CheckedDiv + CheckedMul + PartialOrd + From<u32> + Copy;
}

Chain that has pallet-bridge-messages and dispatch modules.

Associated Types

type Hash: Decode

Hash used in the chain.

type AccountId: Encode + Decode

Accound id on the chain.

type Signer: Encode + Decode

Public key of the chain account that may be used to verify signatures.

type Signature: Encode + Decode

Signature type used on the chain.

type Weight: From<Weight> + PartialOrd

Type of weight that is used on the chain. This would almost always be a regular frame_support::weight::Weight. But since the meaning of weight on different chains may be different, the WeightOf<> construct is used to avoid confusion between different weights.

type Balance: Encode + Decode + CheckedAdd + CheckedDiv + CheckedMul + PartialOrd + From<u32> + Copy

Type of balances that is used on the chain.