Module bridge_runtime_common::refund_relayer_extension

Signed extension that refunds relayer if he has delivered some new messages. It also refunds transaction cost if the transaction is an utility.batchAll() with calls that are: delivering new messsage and all necessary underlying headers (parachain or relay chain).

Structs

ActualFeeRefund

RefundCalculator implementation which refunds the actual transaction fee.

DefaultRefundableParachainId

Default implementation of RefundableParachainId.

PreDispatchData

Data that is crafted in pre_dispatch method and used at post_dispatch.

RefundBridgedParachainMessages

Signed extension that refunds a relayer for new messages coming from a parachain.

RefundableMessagesLane

Default implementation of RefundableMessagesLaneId.

RefundableParachain

Implementation of RefundableParachainId for trait Parachain.

Enums

CallInfo

Type of the call that the extension recognizes.

Traits

RefundCalculator

Refund calculator.

RefundableMessagesLaneId

Trait identifying a bridged messages lane. A relayer might be refunded for delivering messages coming from this lane.

RefundableParachainId

Trait identifying a bridged parachain. A relayer might be refunded for delivering messages coming from this parachain.