Trait bridge_runtime_common::BridgeRuntimeFilterCall

pub trait BridgeRuntimeFilterCall<Call> {
    // Required method
    fn validate(call: &Call) -> TransactionValidity;
}

A duplication of the FilterCall trait.

We need this trait in order to be able to implement it for the messages pallet, since the implementation is done outside of the pallet crate.

Required Methods

fn validate(call: &Call) -> TransactionValidity

Checks if a runtime call is valid.

Object Safety

This trait is not object safe.