Crate bridge_runtime_common

Common types/functions that may be used by runtimes of all bridged chains.

Modules

messages

Types that allow runtime to act as a source/target endpoint of message lanes.

messages_api

Helpers for implementing various message-related runtime API mthods.

messages_call_ext

messages_generation

Helpers for generating message storage proofs, that are used by tests and by benchmarks.

messages_xcm_extension

Module provides utilities for easier XCM handling, e.g: XcmExecutor -> MessageSender -> OutboundMessageQueue | Relayer | XcmRouter <- MessageDispatch <- InboundMessageQueue

priority_calculator

Bridge transaction priority calculator.

refund_relayer_extension

Signed extension that refunds relayer if he has delivered some new messages. It also refunds transaction cost if the transaction is an utility.batchAll() with calls that are: delivering new messsage and all necessary underlying headers (parachain or relay chain).

Macros

generate_bridge_reject_obsolete_headers_and_messages

Declares a runtime-specific BridgeRejectObsoleteHeadersAndMessages signed extension.

Enums

CustomNetworkId

A mapping over NetworkId. Since NetworkId doesn’t include Millau, Rialto and RialtoParachain, we create some synthetic associations between these chains and NetworkId chains.

Traits

BridgeRuntimeFilterCall

A duplication of the FilterCall trait.