Module bridge_runtime_common::extensions

Bridge-specific transaction extensions.

Modules

check_obsolete_extension

Transaction extension that rejects bridge-related transactions, that include obsolete (duplicated) data or do not pass some additional pallet-specific checks.

priority_calculator

Bridge transaction priority calculator.

refund_relayer_extension

Signed extension that refunds relayer if he has delivered some new messages. It also refunds transaction cost if the transaction is an utility.batchAll() with calls that are: delivering new message and all necessary underlying headers (parachain or relay chain).