bridge_runtime_common
Module messages

Types that allow runtime to act as a source/target endpoint of message lanes.

Messages are assumed to be encoded Calls of the target chain. Call-dispatch pallet is used to dispatch incoming messages. Message identified by a tuple of to elements - message lane id and message nonce.

Modules

source

Sub-module that is declaring types required for processing This -> Bridged chain messages.

target

Sub-module that is declaring types required for processing Bridged -> This chain messages.

Traits

BridgedChainWithMessages

Bridged chain that has pallet-bridge-messages module.

MessageBridge

Bidirectional message bridge.

RangeInclusiveExt

A trait defining helper methods for RangeInclusive (start..=end)

ThisChainWithMessages

This chain that has pallet-bridge-messages module.

UnderlyingChainProvider

A trait that provides the type of the underlying chain.

Type Aliases

AccountIdOf

Account id used on the chain.

BalanceOf

Type of balances that is used on the chain.

BridgeMessagesCallOf

The BridgeMessagesCall used by a chain.

BridgedChain

Bridged chain in context of message bridge.

HashOf

Hash used on the chain.

HasherOf

Hasher used on the chain.

ThisChain

This chain in context of message bridge.

UnderlyingChainOf

Underlying chain type.