Module bridge_runtime_common::messages

Types that allow runtime to act as a source/target endpoint of message lanes.

Messages are assumed to be encoded Calls of the target chain. Call-dispatch pallet is used to dispatch incoming messages. Message identified by a tuple of to elements - message lane id and message nonce.

Modules

source

Sub-module that is declaring types required for processing This -> Bridged chain messages.

target

Sub-module that is declaring types required for processing Bridged -> This chain messages.

Structs

MessageTransaction

Message related transaction parameters estimation.

Traits

BridgedChainWithMessages

Bridged chain that has pallet-bridge-messages and dispatch modules.

ChainWithMessages

Chain that has pallet-bridge-messages and dispatch modules.

MessageBridge

Bidirectional message bridge.

ThisChainWithMessages

This chain that has pallet-bridge-messages and dispatch modules.

Functions

transaction_payment

Compute fee of transaction at runtime where regular transaction payment pallet is being used.