bridge_runtime_common::messages
Module source

Sub-module that is declaring types required for processing This -> Bridged chain messages.

Structs

FromBridgedChainMessagesDeliveryProof

Messages delivery proof from bridged chain:

FromThisChainMaximalOutboundPayloadSize

Maximal size of outbound message payload.

FromThisChainMessageVerifier

Message verifier that is doing all basic checks.

TargetHeaderChainAdapter

TargetHeaderChain implementation that is using default types and perform default checks.

Functions

maximal_message_size

Return maximal message size of This -> Bridged chain message.

verify_chain_message

Do basic Bridged-chain specific verification of This -> Bridged chain message.

verify_messages_delivery_proof

Verify proof of This -> Bridged chain messages delivery.

Type Aliases

FromThisChainMessagePayload

Message payload for This -> Bridged chain messages.

ParsedMessagesDeliveryProofFromBridgedChain

‘Parsed’ message delivery proof - inbound lane id and its state.