Module bridge_runtime_common::messages::target

Sub-module that is declaring types required for processing Bridged -> This chain messages.

Structs

FromBridgedChainMessagesProof

Messages proof from bridged chain:

SourceHeaderChainAdapter

SourceHeaderChain implementation that is using default types and perform default checks.

Functions

maximal_incoming_message_dispatch_weight

Return maximal dispatch weight of the message we’re able to receive.

maximal_incoming_message_size

Return maximal message size given maximal extrinsic size.

verify_messages_proof

Verify proof of Bridged -> This chain messages.

Type Aliases

FromBridgedChainMessagePayload

Decoded Bridged -> This message payload.