bridge_runtime_common::messages
Trait MessageBridge

pub trait MessageBridge {
    type ThisChain: ThisChainWithMessages;
    type BridgedChain: BridgedChainWithMessages;
    type BridgedHeaderChain: HeaderChain<UnderlyingChainOf<Self::BridgedChain>>;

    const BRIDGED_MESSAGES_PALLET_NAME: &'static str;
}

Bidirectional message bridge.

Required Associated Constants

const BRIDGED_MESSAGES_PALLET_NAME: &'static str

Name of the paired messages pallet instance at the Bridged chain.

Should be the name that is used in the construct_runtime!() macro.

Required Associated Types

type ThisChain: ThisChainWithMessages

This chain in context of message bridge.

type BridgedChain: BridgedChainWithMessages

Bridged chain in context of message bridge.

type BridgedHeaderChain: HeaderChain<UnderlyingChainOf<Self::BridgedChain>>

Bridged header chain.

Object Safety

This trait is not object safe.