Trait bridge_runtime_common::messages::MessageBridge

pub trait MessageBridge {
    type ThisChain: ThisChainWithMessages;
    type BridgedChain: BridgedChainWithMessages;

    const RELAYER_FEE_PERCENT: u32;
    const THIS_CHAIN_ID: ChainId;
    const BRIDGED_CHAIN_ID: ChainId;
    const BRIDGED_MESSAGES_PALLET_NAME: &'static str;

    fn bridged_balance_to_this_balance(
        bridged_balance: BalanceOf<BridgedChain<Self>>
    ) -> BalanceOf<ThisChain<Self>>;
}

Bidirectional message bridge.

Associated Types

type ThisChain: ThisChainWithMessages

This chain in context of message bridge.

type BridgedChain: BridgedChainWithMessages

Bridged chain in context of message bridge.

Associated Constants

const RELAYER_FEE_PERCENT: u32

Relayer interest (in percents).

const THIS_CHAIN_ID: ChainId

Identifier of this chain.

const BRIDGED_CHAIN_ID: ChainId

Identifier of the Bridged chain.

const BRIDGED_MESSAGES_PALLET_NAME: &'static str

Name of the paired messages pallet instance at the Bridged chain.

Should be the name that is used in the construct_runtime!() macro.

Required methods

fn bridged_balance_to_this_balance(
bridged_balance: BalanceOf<BridgedChain<Self>>
) -> BalanceOf<ThisChain<Self>>

Convert Bridged chain balance into This chain balance.