Trait bridge_runtime_common::messages::ThisChainWithMessages

pub trait ThisChainWithMessages: ChainWithMessages {
    type Call: Encode + Decode;
    fn is_outbound_lane_enabled(lane: &LaneId) -> bool;

    fn maximal_pending_messages_at_outbound_lane() -> MessageNonce;

    fn estimate_delivery_confirmation_transaction(
    ) -> MessageTransaction<Self::Weight>;

    fn transaction_payment(
        transaction: MessageTransaction<Self::Weight>
    ) -> Self::Balance;
}

This chain that has pallet-bridge-messages and dispatch modules.

Associated Types

type Call: Encode + Decode

Call type on the chain.

Loading content...
Required methods

fn is_outbound_lane_enabled(lane: &LaneId) -> bool

Are we accepting any messages to the given lane?

fn maximal_pending_messages_at_outbound_lane() -> MessageNonce

Maximal number of pending (not yet delivered) messages at This chain.

Any messages over this limit, will be rejected.

fn estimate_delivery_confirmation_transaction(
) -> MessageTransaction<Self::Weight>

Estimate size and weight of single message delivery confirmation transaction at This chain.

fn transaction_payment(
transaction: MessageTransaction<Self::Weight>
) -> Self::Balance

Returns minimal transaction fee that must be paid for given transaction at This chain.

Loading content...