bridge_runtime_common
Module messages_call_ext

Structs

BaseMessagesProofInfo

Generic info about a messages delivery/confirmation proof.

CallHelper

Helper struct that provides methods for working with a call supported by CallInfo.

ReceiveMessagesDeliveryProofInfo

Info about a ReceiveMessagesDeliveryProof call which tries to update a single lane.

ReceiveMessagesProofInfo

Info about a ReceiveMessagesProof call which tries to update a single lane.

UnrewardedRelayerOccupation

Occupation state of the unrewarded relayers vector.

Enums

CallInfo

Info about a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call which tries to update a single lane.

Traits

MessagesCallSubType

Trait representing a call that is a sub type of pallet_bridge_messages::Call.