Trait bridge_runtime_common::messages_call_ext::MessagesCallSubType

pub trait MessagesCallSubType<T: Config<I, RuntimeCall = Self>, I: 'static>: IsSubType<CallableCallFor<Pallet<T, I>, T>> {
    // Required methods
    fn receive_messages_proof_info(&self) -> Option<ReceiveMessagesProofInfo>;
    fn receive_messages_delivery_proof_info(
        &self
    ) -> Option<ReceiveMessagesDeliveryProofInfo>;
    fn call_info(&self) -> Option<CallInfo>;
    fn call_info_for(&self, lane_id: LaneId) -> Option<CallInfo>;
    fn check_obsolete_call(&self) -> TransactionValidity;
}

Trait representing a call that is a sub type of pallet_bridge_messages::Call.

Required Methods

fn receive_messages_proof_info(&self) -> Option<ReceiveMessagesProofInfo>

Create a new instance of ReceiveMessagesProofInfo from a ReceiveMessagesProof call.

fn receive_messages_delivery_proof_info( &self ) -> Option<ReceiveMessagesDeliveryProofInfo>

Create a new instance of ReceiveMessagesDeliveryProofInfo from a ReceiveMessagesDeliveryProof call.

fn call_info(&self) -> Option<CallInfo>

Create a new instance of CallInfo from a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call.

fn call_info_for(&self, lane_id: LaneId) -> Option<CallInfo>

Create a new instance of CallInfo from a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call, if the call is for the provided lane.

fn check_obsolete_call(&self) -> TransactionValidity

Check that a ReceiveMessagesProof or a ReceiveMessagesDeliveryProof call is trying to deliver/confirm at least some messages that are better than the ones we know of.