bridge_runtime_common
Module messages_generation

Helpers for generating message storage proofs, that are used by tests and by benchmarks.

Functions

encode_all_messages

Simple and correct message data encode function.

encode_lane_data

Simple and correct outbound lane data encode function.

grow_trie_leaf_value

Add extra data to the trie leaf value so that it’ll be of given size.

prepare_messages_storage_proof

Prepare storage proof of given messages.