Module bridge_runtime_common::messages_xcm_extension

Module provides utilities for easier XCM handling, e.g: XcmExecutor -> MessageSender -> OutboundMessageQueue | Relayer | XcmRouter <- MessageDispatch <- InboundMessageQueue

Structs

LocalXcmQueueManager

Manager of local XCM queues (and indirectly - underlying transport channels) that controls the queue state.

SenderAndLane

A pair of sending chain location and message lane, used by this chain to send messages over the bridge.

XcmBlobHaulerAdapter

XCM bridge adapter which connects XcmBlobHauler with pallet_bridge_messages and makes sure that XCM blob is sent to the outbound lane to be relayed.

XcmBlobMessageDispatch

XcmBlobMessageDispatch is responsible for dispatching received messages

Enums

XcmBlobMessageDispatchResult

Message dispatch result type for single message

Traits

XcmBlobHauler

XcmBlobHauler is responsible for sending messages to the bridge “point-to-point link” from one side, where on the other it can be dispatched by XcmBlobMessageDispatch.

Type Definitions

XcmAsPlainPayload

Plain “XCM” payload, which we transfer through bridge