bridge_runtime_common::messages_xcm_extension
Trait XcmBlobHauler

pub trait XcmBlobHauler {
    type Runtime: MessagesConfig<Self::MessagesInstance>;
    type MessagesInstance: 'static;
    type SenderAndLane: Get<SenderAndLane>;
    type ToSourceChainSender: SendXcm;
    type CongestedMessage: Get<Option<Xcm<()>>>;
    type UncongestedMessage: Get<Option<Xcm<()>>>;

    // Provided method
    fn supports_congestion_detection() -> bool { ... }
}

XcmBlobHauler is responsible for sending messages to the bridge “point-to-point link” from one side, where on the other it can be dispatched by XcmBlobMessageDispatch.

Required Associated Types

type Runtime: MessagesConfig<Self::MessagesInstance>

Runtime that has messages pallet deployed.

type MessagesInstance: 'static

Instance of the messages pallet that is used to send messages.

type SenderAndLane: Get<SenderAndLane>

Returns lane used by this hauler.

type ToSourceChainSender: SendXcm

Actual XCM message sender (HRMP or UMP) to the source chain location (Self::SenderAndLane::get().location).

type CongestedMessage: Get<Option<Xcm<()>>>

An XCM message that is sent to the sending chain when the bridge queue becomes congested.

type UncongestedMessage: Get<Option<Xcm<()>>>

An XCM message that is sent to the sending chain when the bridge queue becomes not congested.

Provided Methods

fn supports_congestion_detection() -> bool

Returns true if we want to handle congestion.

Object Safety

This trait is not object safe.